refactor: 🚚 move Config
