+++ src/main/java/com/zhanganzhi/chathub/core/adaptor/AbstractAdaptor.java
@@ -1,7 +1,7 @@
 package com.zhanganzhi.chathub.core.adaptor;
 
 import com.zhanganzhi.chathub.ChatHub;
-import com.zhanganzhi.chathub.core.Config;
+import com.zhanganzhi.chathub.core.config.Config;
 import com.zhanganzhi.chathub.core.events.MessageEvent;
 import com.zhanganzhi.chathub.core.events.ServerChangeEvent;
 import com.zhanganzhi.chathub.core.formatter.IFormatter;
