+++ src/main/java/com/zhanganzhi/chathub/platforms/discord/DiscordReceiver.java
@@ -1,7 +1,7 @@
 package com.zhanganzhi.chathub.platforms.discord;
 
 import com.zhanganzhi.chathub.ChatHub;
-import com.zhanganzhi.chathub.core.Config;
+import com.zhanganzhi.chathub.core.config.Config;
 import com.zhanganzhi.chathub.core.EventHub;
 import com.zhanganzhi.chathub.core.events.MessageEvent;
 import com.zhanganzhi.chathub.platforms.Platform;
